+++ legadilo/website/urls.py
@@ -14,4 +14,5 @@ urlpatterns = [
     path("privacy/", TemplateView.as_view(template_name="website/privacy.html"), name="privacy"),
     path("manifest.json", views.manifest_view, name="manifest"),
     path("favicon.ico", views.default_favicon_view, name="favicon"),
+    path(".well-known/security.txt", views.security_txt_view, name="security_txt"),
 ]